FIXED IT!
Functionality added in to handle The Stat Evolvers and preReqs>

To-do:

Check the preReq shares a name of one of the other jobs. If it does and
it is not the class being edited, then it goes through. If it matches
any of the following, it sends an alert to the user.